ndk/media: Move `MediaFormat` to its own module
This type is used outside of `media_codec`, for example in the
`media_extractor` bindings.
